refactor(api): migrate badge api scope
