+++ api/lib/infrastructure/repositories/target-profile-repository.js
@@ -1,8 +1,8 @@
 import _ from 'lodash';
 
 import { knex } from '../../../db/knex-database-connection.js';
+import { Badge } from '../../../src/evaluation/domain/models/Badge.js';
 import { NotFoundError } from '../../../src/shared/domain/errors.js';
-import { Badge } from '../../../src/shared/domain/models/Badge.js';
 import { ObjectValidationError } from '../../domain/errors.js';
 import { TargetProfile } from '../../domain/models/index.js';
 import { DomainTransaction } from '../DomainTransaction.js';
