+++ api/src/evaluation/application/badges/badges-controller.js
@@ -1,11 +1,12 @@
-import { sharedUsecases } from '../../domain/usecases/index.js';
+import { evaluationUsecases as usecases } from '../../../evaluation/domain/usecases/index.js';
+import { sharedUsecases } from '../../../shared/domain/usecases/index.js';
 import * as badgeSerializer from '../../infrastructure/serializers/jsonapi/badge-serializer.js';
 
 const updateBadge = async function (request, h) {
   const badgeId = request.params.id;
   const badge = badgeSerializer.deserialize(request.payload);
 
-  const updatedBadge = await sharedUsecases.updateBadge({ badgeId, badge });
+  const updatedBadge = await usecases.updateBadge({ badgeId, badge });
 
   return h.response(badgeSerializer.serialize(updatedBadge)).code(204);
 };
