+++ api/src/evaluation/application/badges/index.js
@@ -1,7 +1,7 @@
 import Joi from 'joi';
 
-import { securityPreHandlers } from '../../application/security-pre-handlers.js';
-import { identifiersType } from '../../domain/types/identifiers-type.js';
+import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
+import { identifiersType } from '../../../shared/domain/types/identifiers-type.js';
 import { badgesController } from './badges-controller.js';
 
 const register = async function (server) {
