+++ api/src/evaluation/routes.js
@@ -1,6 +1,7 @@
 import * as answersRoutes from './application/answers/index.js';
 import * as autonomousCoursesRoutes from './application/autonomous-courses/index.js';
 import * as badgeCriteriaRoutes from './application/badge-criteria/index.js';
+import * as badgesRoutes from './application/badges/index.js';
 import * as competenceEvaluationsRoutes from './application/competence-evaluations/index.js';
 import * as feedbacksRoutes from './application/feedbacks/index.js';
 import * as progressionsRoutes from './application/progressions/index.js';
@@ -10,15 +11,16 @@ import * as stageCollectionRoutes from './application/stage-collections/index.js
 import * as stagesRoutes from './application/stages/index.js';
 
 const evaluationRoutes = [
-  smartRandomSimulatorRoutes,
   answersRoutes,
   autonomousCoursesRoutes,
   badgeCriteriaRoutes,
+  badgesRoutes,
   competenceEvaluationsRoutes,
   stageCollectionRoutes,
   feedbacksRoutes,
   progressionsRoutes,
   scorecardsRoutes,
+  smartRandomSimulatorRoutes,
   stagesRoutes,
 ];
