+++ api/src/prescription/campaign/infrastructure/repositories/target-profile-repository.js
@@ -1,6 +1,6 @@
 import { knex } from '../../../../../db/knex-database-connection.js';
 import { TargetProfile } from '../../../../../lib/domain/models/index.js';
-import { Badge } from '../../../../shared/domain/models/Badge.js';
+import { Badge } from '../../../../evaluation/domain/models/Badge.js';
 
 const getByCampaignId = async function (campaignId) {
   const { targetProfileId } = await knex('campaigns').select('targetProfileId').where({ id: campaignId }).first();
