+++ api/src/shared/routes.js
@@ -1,7 +1,6 @@
 import * as assessmentsRoutes from './application/assessments/index.js';
-import * as badgesRoutes from './application/badges/index.js';
 import * as challengesRoutes from './application/challenges/index.js';
 
-const sharedRoutes = [assessmentsRoutes, badgesRoutes, challengesRoutes];
+const sharedRoutes = [assessmentsRoutes, challengesRoutes];
 
 export { sharedRoutes };
