+++ api/tests/evaluation/integration/domain/usecases/create-badge_test.js
@@ -1,15 +1,15 @@
 import _ from 'lodash';
 
 import * as targetProfileRepository from '../../../../../lib/infrastructure/repositories/target-profile-repository.js';
+import { Badge } from '../../../../../src/evaluation/domain/models/Badge.js';
+import { createBadge } from '../../../../../src/evaluation/domain/usecases/create-badge.js';
 import * as badgeCriteriaRepository from '../../../../../src/evaluation/infrastructure/repositories/badge-criteria-repository.js';
+import * as badgeRepository from '../../../../../src/evaluation/infrastructure/repositories/badge-repository.js';
 import {
   AlreadyExistingEntityError,
   MissingBadgeCriterionError,
   NotFoundError,
 } from '../../../../../src/shared/domain/errors.js';
-import { Badge } from '../../../../../src/shared/domain/models/Badge.js';
-import { createBadge } from '../../../../../src/shared/domain/usecases/create-badge.js';
-import * as badgeRepository from '../../../../../src/shared/infrastructure/repositories/badge-repository.js';
 import { catchErr, databaseBuilder, expect, knex, mockLearningContent, sinon } from '../../../../test-helper.js';
 
 describe('Integration | UseCases | create-badge', function () {
