+++ api/tests/evaluation/integration/infrastructure/repositories/badge-repository_test.js
@@ -1,8 +1,8 @@
 import omit from 'lodash/omit.js';
 
+import { Badge } from '../../../../../src/evaluation/domain/models/Badge.js';
+import * as badgeRepository from '../../../../../src/evaluation/infrastructure/repositories/badge-repository.js';
 import { AlreadyExistingEntityError } from '../../../../../src/shared/domain/errors.js';
-import { Badge } from '../../../../../src/shared/domain/models/Badge.js';
-import * as badgeRepository from '../../../../../src/shared/infrastructure/repositories/badge-repository.js';
 import { catchErr, databaseBuilder, expect, knex } from '../../../../test-helper.js';
 
 describe('Integration | Repository | Badge', function () {
