+++ api/tests/evaluation/unit/infrastructure/serializers/jsonapi/badge-creation-serializer_test.js
@@ -1,5 +1,5 @@
-import { deserializer } from '../../../../../lib/infrastructure/serializers/jsonapi/badge-creation-serializer.js';
-import { expect } from '../../../../test-helper.js';
+import { deserializer } from '../../../../../../src/evaluation/infrastructure/serializers/jsonapi/badge-creation-serializer.js';
+import { expect } from '../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | badge-creation-serializer', function () {
   describe('#deserialize', function () {
