+++ api/tests/prescription/campaign/integration/domain/usecases/get-campaign_test.js
@@ -1,7 +1,7 @@
 import * as campaignRepository from '../../../../../../lib/infrastructure/repositories/campaign-repository.js';
+import * as badgeRepository from '../../../../../../src/evaluation/infrastructure/repositories/badge-repository.js';
 import { usecases } from '../../../../../../src/prescription/campaign/domain/usecases/index.js';
 import * as campaignReportRepository from '../../../../../../src/prescription/campaign/infrastructure/repositories/campaign-report-repository.js';
-import * as badgeRepository from '../../../../../../src/shared/infrastructure/repositories/badge-repository.js';
 import { databaseBuilder, expect, mockLearningContent } from '../../../../../test-helper.js';
 
 describe('Integration | UseCase | get-campaign', function () {
